split component into several parts
